+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Unreleased
 ----------
 - Added `unsable-attach-point` lint
 - Added `builtin_lints` function for retrieving list of built-in lints
+- Added support for disabling lints via code comments of the form
+  `bpflint: disable=<lint-name>`
 - Ensured `lint` reports matches in source code order
 - Fixed `report_terminal` to correctly handle matches on top most line